requirements.txt
torch
torchvision
timm
flask
Pillow
scikit-learn
matplotlib
seaborn


